SMART_PORTAL/backend/requirements.txt
Flask
flask_jwt_extended
flask_bcrypt
flask_cors
pymongo
dnspython
python-dotenv
pypdf
python-docx
openpyxl
reportlab
